Moving the flake8 visitors to a single entry point in main.py. Currently moved send_file and secure_set_cookies under the new entrypoint.
